Fix audit point 22: use long nanoseconds internally for timing precision
- Time.getTime() returns double instead of float, eliminating cast-to-float
  precision loss on the absolute timestamp
- Game loop tracks beginNanos/deltaNanos/totalNanos as long via
  System.nanoTime(); float cast happens only at GameTime.set() where
  elapsedSeconds is always a small number (~16ms) with no precision risk
- GameTime public API stays float as specified in the audit

+++ core/src/main/java/hu/mudlee/core/Time.java
@@ -3,7 +3,7 @@ package hu.mudlee.core;
 public class Time {
     public static long timeStarted = System.nanoTime();
 
-    public static float getTime() {
-        return (float) ((System.nanoTime() - timeStarted) * 1E-9);
+    public static double getTime() {
+        return (System.nanoTime() - timeStarted) * 1e-9;
     }
 }
